moved model to a separate file
