+++ libs/designer-ui/src/lib/templates/templatesSection.tsx
@@ -1,80 +1,6 @@
-import { Dropdown, Field, Input, Label, Link, Option, Radio, RadioGroup, type Slot, Text } from '@fluentui/react-components';
+import { Dropdown, Field, Input, Label, Link, Option, Radio, RadioGroup, Text } from '@fluentui/react-components';
 import { Open16Regular } from '@fluentui/react-icons';
-
-interface BaseTemplatesSectionItem {
-  type: 'text' | 'textField' | 'dropdown' | 'radioGroup';
-  label?: string | React.ReactNode;
-  value: string | undefined;
-}
-
-interface TextItem extends BaseTemplatesSectionItem {
-  type: 'text';
-}
-
-interface BaseFieldItem extends BaseTemplatesSectionItem {
-  id?: string;
-  required?: boolean;
-  disabled?: boolean;
-  errorMessage?: string;
-  hint?: string;
-}
-
-interface BaseTextInputFieldItem extends BaseFieldItem {
-  placeholder?: string;
-  onBlur?: () => Promise<void>;
-}
-
-interface TextFieldItem extends BaseTextInputFieldItem {
-  type: 'textField';
-  onChange: (value: string) => void;
-}
-
-interface DropdownItem extends BaseTextInputFieldItem {
-  type: 'dropdown';
-  options: {
-    id: string;
-    label: string;
-    value: string;
-  }[];
-  onOptionSelect: (selectedOptions: string[]) => void;
-  selectedOptions: string[];
-  multiselect?: boolean;
-}
-
-interface RadioGroupItem extends BaseFieldItem {
-  type: 'radioGroup';
-  onOptionSelect: (selectedValue: string) => void;
-  options: {
-    id: string;
-    label: Slot<typeof Label>;
-    value: string;
-  }[];
-}
-
-export type TemplatesSectionItem = TextItem | TextFieldItem | DropdownItem | RadioGroupItem;
-
-export interface BaseTemplatesSectionProps {
-  title: string;
-  isTitleRequired?: boolean;
-  titleHtmlFor?: string;
-  description?: string;
-  descriptionLink?: {
-    text: string;
-    href: string;
-  };
-}
-
-export interface ContentBasedProps extends BaseTemplatesSectionProps {
-  items: TemplatesSectionItem[];
-  children?: never;
-}
-
-export interface ChildrenBasedProps extends BaseTemplatesSectionProps {
-  children: React.ReactNode;
-  items?: never;
-}
-
-export type TemplatesSectionProps = ContentBasedProps | ChildrenBasedProps;
+import type { TemplatesSectionItem, TemplatesSectionProps } from './templatesSectionModel';
 
 export const TemplatesSection = ({
   title,
